runqemu: don't use qemu -vga vmware when using OVMF BIOS
The OVMF BIOS seems to prefer the defautl Cirrus emulated VGA adaptor,
otherwise the QEMU window just shows a blank screen.

NOTE: OVMF is detected simply by looking for the substring "ovmf" in the
biosdir argument to runqemu. This is, of course, not a robust test, but it
works well enough if you're just using the ovmf recipe in meta-luv.

Signed-off-by: Max Eliaser <max.eliaser@intel.com>