meta-luv: ovmf: generate 32-bit firmware images when appropriate
This fixes an issue where the OVMF recipe always built an x86_64 firmware
regardless of the target architecture or target machine. This issue manifested
itself when qemu-system-i386 was used on an OVMF firmware image-- QEMU would
exit immediately rather than proceeding with the boot process.

The x86_64 firmware was unaffected by this issue.

Signed-off-by: Max Eliaser <max.eliaser@intel.com>